requirements.txt
opencv-python==4.5.5.64
numpy==1.21.5
Pillow==9.0.1
pytesseract==0.3.9
openai==0.27.0
python-dotenv==0.19.2
Flask==2.0.1
Werkzeug==2.0.1